+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
 
 import '../../sentry_flutter.dart';
+import '../event_processor/flutter_enricher_event_processor.dart';
 import '../native/sentry_native.dart';
 
 /// This key must be used so that the web interface displays the events nicely
@@ -21,6 +23,9 @@ typedef AdditionalInfoExtractor = Map<String, dynamic>? Function(
 /// For example, if the application starts, there is no previous route.
 /// The [RouteSettings] is null if a developer has not specified any
 /// RouteSettings.
+///
+/// The current route name will also be set to [SentryEvent]
+/// `contexts.app.view_names` by [FlutterEnricherEventProcessor].
 ///
 /// [SentryNavigatorObserver] must be added to the [navigation observer](https://api.flutter.dev/flutter/material/MaterialApp/navigatorObservers.html) of
 /// your used app. This is an example for [MaterialApp](https://api.flutter.dev/flutter/material/MaterialApp/navigatorObservers.html),
@@ -84,11 +89,17 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
 
   ISentrySpan? _transaction;
 
+  static String? _currentRouteName;
+
+  @internal
+  static String? get currentRouteName => _currentRouteName;
+
   @override
   void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
     super.didPush(route, previousRoute);
 
-    _setCurrentRoute(route);
+    _setCurrentRouteName(route);
+    _setCurrentRouteNameAsTransaction(route);
 
     _addBreadcrumb(
       type: 'didPush',
@@ -104,7 +115,9 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
   void didReplace({Route<dynamic>? newRoute, Route<dynamic>? oldRoute}) {
     super.didReplace(newRoute: newRoute, oldRoute: oldRoute);
 
-    _setCurrentRoute(newRoute);
+    _setCurrentRouteName(newRoute);
+    _setCurrentRouteNameAsTransaction(newRoute);
+
     _addBreadcrumb(
       type: 'didReplace',
       from: oldRoute?.settings,
@@ -116,7 +129,9 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
   void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
     super.didPop(route, previousRoute);
 
-    _setCurrentRoute(previousRoute);
+    _setCurrentRouteName(previousRoute);
+    _setCurrentRouteNameAsTransaction(previousRoute);
+
     _addBreadcrumb(
       type: 'didPop',
       from: route.settings,
@@ -147,7 +162,11 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
         ?.name;
   }
 
-  Future<void> _setCurrentRoute(Route<dynamic>? route) async {
+  Future<void> _setCurrentRouteName(Route<dynamic>? route) async {
+    _currentRouteName = _getRouteName(route);
+  }
+
+  Future<void> _setCurrentRouteNameAsTransaction(Route<dynamic>? route) async {
     final name = _getRouteName(route);
     if (name == null) {
       return;