+++ sqflite/lib/src/utils/sentry_database_span_attributes.dart
@@ -10,3 +10,12 @@ void setDatabaseAttributeData(ISentrySpan? span, String? dbName) {
     span?.setData(SentryDatabase.dbNameKey, dbName);
   }
 }
+
+/// Sets the database attributes on the [breadcrumb].
+/// It contains the database system and the database name.
+void setDatabaseAttributeOnBreadcrumb(Breadcrumb breadcrumb, String? dbName) {
+  breadcrumb.data?[SentryDatabase.dbSystemKey] = SentryDatabase.dbSystem;
+  if (dbName != null) {
+    breadcrumb.data?[SentryDatabase.dbNameKey] = dbName;
+  }
+}